feat: floating button 공통 컴포넌트로 구현
